Moving PlayoffDetails to a component
The code for the Playoff Details was duplicated on the Bottom Buttons and the Cheat Sheet pages. I have consolidated these two into one component.
