+++ .vscode/launch.json
@@ -4,7 +4,7 @@
     // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [
-        {
+    {
             "type": "firefox",
             "request": "launch",
             "reAttach": true,
